+++ client/static/components/js/image/image.js
@@ -5,41 +5,60 @@ ImageComponent = {
     }
 };
 
-var ImageViewer = function(sandbox){
+const ImageViewer = function (sandbox) {
     this.container = '#' + sandbox.container;
     this.sandbox = sandbox;
 
-    function toDataUrl(url, callback) {
-        var xhr = new XMLHttpRequest();
-        xhr.responseType = 'blob';
-        xhr.onload = function() {
-            var reader = new FileReader();
-            reader.onloadend = function() {
-                callback(reader.result);
-            };
-            reader.readAsDataURL(xhr.response);
-        };
-        xhr.open('GET', url);
-        xhr.send();
-    }
-
     // ---- window interface -----
-    this.receiveData = function(data) {
-        return new Promise(resolve => {
-            $(this.container).empty();
-            $(this.container).append('<img src="' + data + '" style="width: 100%; height: 100%;"></img>');
-
-            resolve();
-        });
+    this.receiveData = function (mimeType, data) {
+        $(this.container).empty();
+        $(this.container).append('<img src="data:' + mimeType + ';base64,' + data + '" style="width: 100%; height: 100%;"></img>');
     };
 
-    var self = this;
+    const getMimeType = async function (link) {
+        const FORMAT = "_format";
+        const MIME_TYPE = "_mime_type";
+
+        const template = new sc.ScTemplate();
+        template.tripleWithRelation(
+            new sc.ScAddr(link),
+            sc.ScType.EdgeDCommonVar,
+            [sc.ScType.NodeVar, FORMAT],
+            sc.ScType.EdgeAccessVarPosPerm,
+            new sc.ScAddr(window.scKeynodes["nrel_format"]),
+        );
+        template.tripleWithRelation(
+            FORMAT,
+            sc.ScType.EdgeDCommonVar,
+            [sc.ScType.NodeVar, MIME_TYPE],
+            sc.ScType.EdgeAccessVarPosPerm,
+            new sc.ScAddr(window.scKeynodes["nrel_mimetype"]),
+        );
+        const result = await scClient.templateSearch(template);
+        if (result.length) {
+            const mimeLink = result[0].get(MIME_TYPE);
+            const contents = await scClient.getLinkContents([mimeLink]);
+
+            if (contents.length) {
+                return contents[0].data;
+            }
+
+            return "";
+        }
+
+        return "";
+    }
+
     if (this.sandbox.addr) {
-        toDataUrl('api/link/content/?addr=' + this.sandbox.addr, function(base64Img) {
-            self.receiveData(base64Img);
+        window.scClient.getLinkContents([new sc.ScAddr(this.sandbox.addr)]).then((contents) => {
+            if (contents.length) {
+                let base64 = contents[0].data;
+                getMimeType(this.sandbox.addr).then((mimeType) => {
+                    this.receiveData(mimeType, base64);
+                });
+            }
         });
     }
 };
-
 
 SCWeb.core.ComponentManager.appendComponentInitialize(ImageComponent);