+++ docs/changelog.md
@@ -5,13 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.7.0-Rebirth] - 12.10.2022
 ### Added
+- Add images mime types handling from kb
 - Add ctrl+c logic on py-server
 - Add logging levels
 - Add search by content substring through sc-server
 - Add latex docs from OSTIS Standard
 
 ### Changed
+- Move search by substring from py-server to js-server
+- Move links/contents handling from py-server to js-server
 - Switch on py-sc-client on py-server and ts-sc-client in js-server
 - Rename `SCsComponent` to `SCnComponent`
 - Change external language from `scs_code` to `scn_code` for `SCsComponent`